AI Tutor/requirements.txt
flask==3.1.3
langchain==0.3.18
langchain-groq==0.2.2
langgraph==0.2.60
python-dotenv==1.0.0
